% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{annualise_data}
\alias{annualise_data}
\title{Annualise monthly data}
\usage{
annualise_data(data)
}
\arguments{
\item{data}{Tidy long form data to be reduced to annual resolution.}
}
\value{
Annual data.
}
\description{
Annualise monthly data
}
