% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{change_baseline}
\alias{change_baseline}
\title{Title}
\usage{
change_baseline(data, baseline)
}
\arguments{
\item{data}{Tidy dataset}

\item{baseline}{Baseline start and end year in a vector}
}
\value{
Tidy dataset with selected baseline
}
\description{
Title
}
