% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{pad_to_full_year}
\alias{pad_to_full_year}
\title{Pad empty month rows to the end of last year}
\usage{
pad_to_full_year(data)
}
\arguments{
\item{data}{Data to be padded to a full year.}
}
\value{
Data padded to full year with NAs as necessary.
}
\description{
Pad empty month rows to the end of last year
}
