% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_berkeley.R
\name{read_berkeley}
\alias{read_berkeley}
\title{Read Berkeley Earth monthly data in tidy form}
\usage{
read_berkeley()
}
\value{
Tidy Berkeley Earth time series
}
\description{
Read Berkeley Earth monthly data in tidy form
}
