% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_climate_ts.R
\name{read_climate_ts}
\alias{read_climate_ts}
\title{Download and read climate time series}
\usage{
read_climate_ts(
  name = "hadcrut",
  resolution = "monthly",
  pad_last_year = FALSE,
  baseline = c(1881, 1920)
)
}
\arguments{
\item{name}{One of "hadcrut", "gistemp", "noaa", "berkeley".}

\item{resolution}{Either "monthly" or "annual".}

\item{pad_last_year}{Pad months to the end of last year with NAs.
If resolution is annual, last partial year is averaged and included in the data.}

\item{baseline}{Baseline start and end year in a vector.}
}
\value{
Tidy, long form climate time series data set.
}
\description{
A helper function to load up-to-date common climate timeseries
data sets, and to format them in consistently named tidy long form data.
}
