% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hadcrut5.R
\name{read_hadcrut5}
\alias{read_hadcrut5}
\title{Read HadCRUT5 monthly data in tidy form}
\usage{
read_hadcrut5()
}
\value{
A tidy HadCRUT5 data set
}
\description{
Read HadCRUT5 monthly data in tidy form
}
