% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_noaa.R
\name{read_noaa}
\alias{read_noaa}
\title{Read NOAA GlobalTemp v5 monthly data in tidy form}
\usage{
read_noaa()
}
\value{
Tidy long form NOAA time series data.
}
\description{
Read NOAA GlobalTemp v5 monthly data in tidy form
}
