% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{scrape_noaa_url}
\alias{scrape_noaa_url}
\title{Scrape NOAA url from NOAA timeseries directory}
\usage{
scrape_noaa_url()
}
\value{
Url to current NOAA GlobalTemp time series file.
}
\description{
Scrape NOAA url from NOAA timeseries directory
}
