% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_climate_ts.R
\name{read_climate_ts}
\alias{read_climate_ts}
\title{Read climate time series in tidy long form}
\usage{
read_climate_ts(
  name = "hadcrut",
  resolution = "monthly",
  pad_last_year = FALSE,
  baseline = c(1881, 1920)
)
}
\arguments{
\item{name}{One of "hadcrut", "gistemp", "noaa", "berkeley" ...TODO}

\item{resolution}{Either "monthly" or "annual".}

\item{pad_last_year}{Pad months to the end of last year with NAs.}

\item{baseline}{Baseline start and end year in a vector.}
}
\value{
Tidy, long form climate time series data set.
}
\description{
Read climate time series in tidy long form
}
