% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_climate_ts.R
\name{read_climate_ts}
\alias{read_climate_ts}
\title{Read climate time series in tidy long form}
\usage{
read_climate_ts(
  name = "hadcrut5",
  resolution = "monthly",
  pad_last_year = FALSE
)
}
\arguments{
\item{name}{One of "hadcrut5", "gistemp", others ...TODO}

\item{resolution}{Either "monthly" or "annual".}

\item{pad_last_year}{Pad months to the end of last year with NAs.}
}
\value{
Tidy climate data set
}
\description{
Read climate time series in tidy long form
}
